requirements.txt
streamlit
openai
python-dotenv
PyPDF2
pdfminer.six
docx2txt
pandas
